feat(webportal): split into separate file
also adds support for attachments and removes the `web_asset_store.xml` component of the asset server URL
