+++ specifyweb/backend/stored_queries/tests/test_views/test_export_web_portal.py
@@ -1,4 +1,4 @@
-from unittest.mock import Mock, patch
+from unittest.mock import MagicMock, Mock, patch
 
 from django.test import Client
 
@@ -24,3 +24,46 @@ class TestExportWebPortal(SQLAlchemySetup):
         self.assertTrue(thread.return_value.daemon)
         thread.return_value.start.assert_called_once()
         self._assertContentEqual(response, "OK")
+
+    def test_portal_attachment_map(self):
+        from specifyweb.backend.stored_queries import execution
+
+        class FakeAttachment:
+            id = 5291
+            attachmentlocation = "sp6896513492722436219.att.JPG"
+            origfilename = "29432.JPG"
+
+        class FakeJoinRecord:
+            collectionobject_id = 123
+            attachment = FakeAttachment()
+
+        class FakeJoinQuery:
+            def select_related(self, *_args, **_kwargs):
+                return [FakeJoinRecord()]
+
+        class FakeJoinManager:
+            def __init__(self):
+                self.filter_kwargs = None
+
+            def filter(self, **kwargs):
+                self.filter_kwargs = kwargs
+                return FakeJoinQuery()
+
+        fake_join_manager = FakeJoinManager()
+        fake_base_model = type("Collectionobject", (), {"_meta": MagicMock(app_label="specifyweb")})
+        fake_table = MagicMock()
+        fake_table.attachments_field = MagicMock()
+
+        with patch.object(execution.datamodel, "get_table_by_id", return_value=fake_table), patch.object(
+            execution, "get_model_by_table_id", return_value=fake_base_model
+        ), patch.object(execution.apps, "get_model", return_value=type("Collectionobjectattachment", (), {"objects": fake_join_manager})):
+            result = execution._portal_attachment_map(1, [123])
+
+        self.assertEqual(
+            fake_join_manager.filter_kwargs,
+            {"collectionobject_id__in": [123]},
+        )
+        self.assertEqual(
+            result["123"],
+            '[{AttachmentID:5291,AttachmentLocation:"sp6896513492722436219.att.JPG",Title:"29432.JPG"}]',
+        )
